Use typescript for OutlineKeySettings
